Add files via upload
Files to transform original data. The scripts will sum the data by precinct, and pivot the data around the candidates and vote type. These are preliminary. The original work was done manually. As a way to reproduce a lot of the charts, some of these will be done using scripts for transparency. Work in Progress.